Centralizado de definicion de tipos y funciones de fechas
Agregadas utilidades y tipos compartidos para turnos y disponibilidad, y refactorizados los componentes para utilizarlos
